More refactoring
Got it to work without maps, but had to chop it up into different consts
